--configure now supports per-bucket checks
Previously, --configure would perform an access check by trying to list
all buckets for the account. This requires the S3 ListAllMyBuckets
permission which is typically not available to delegated IAM accounts.
With this change, --configure now accepts an (optional) bucket uri as a
parameter and if it's provided, the access check will just verify
access to this bucket individually.

i.e.

s3cmd --configure  # Access Denied if the account lacks ListAllMyBuckets

But

s3cmd --configure s3://some-bucket # Still work