* Force MIME type with --mime-type=abc/xyz
Option --mime-type now has a different meaning - from now on
it forces setting the given MIME type instead of setting a default
MIME type. For changing the default from binary/octet-stream use
a new option --default-mime-type.
Also --guess-mime-type is no longer on by default and if set to true
in the config file can be turned back off with --no-guess-mime-type