add local tree MD5 caching
This creates and maintains a cache (aka HashCache) of each inode in
the local tree.  This is used to avoid doing local disk I/O to
calculate an MD5 value for a file if it's inode, mtime, and size
haven't changed.  If these values have changed, then it does the disk
I/O.

This introduces command line option --cache-file <foo>.  The file is
created if it does not exist, is read upon start and written upon
close. The contents are only useful for a given directory tree, so
caches should not be reused for different directory tree syncs.