sync: refactor parent/child and single process code
os.fork() and os.wait() don't exist on Windows, and the
multiprocessing module doesn't exist until python 2.6.  So instead, we
conditionalize calling os.fork() depending on its existance, and on
there being > 1 destination.

Also simply rearranges the code so that subfunctions within
local2remote are defined at the top of their respective functions, for
better readability through the main execution of the function.