fix --cf-invalidate after merge
The --cf-invalidate code path in sync_local2remote would traceback
from undefined variables.

Refactor this code path into a subfunction, saving necessary variables
earlier in execution.  This only invalidates the paths based on the
last uploaded destination in the list (if there are multiple
destinations).  If this isn't the desired behavior, don't upload to
multiple destinations _and_ try to invalidate on CF simultaneously.