add --max-delete for del and sync commands, fixes bug #155
Fixes https://github.com/s3tools/s3cmd/issues/155

Similar to rsync --max-delete, allow the user to specify a maximum
number of objects to delete for the [del] and [sync] commands.  If
that number would be exceeded, output a warning message and do not
delete anything.  This prevents a simple invocation error
(e.g. launching from the wrong directory) from erasing all the data in
a bucket accidentally.

By default, this is disabled.  A user must explicitly specify
--max-delete=NUM for this protection to be available, just as in
rsync.