Don't emit "cowardly" message unless we have source files & delete files
We were looking at the wrong (post-compare) list lengths, rather than
the original (discovered by walk) list lengths.  And we weren't
checking to see if we had any files to delete, or not, before issuing
the "cowardly" message.  So for a "null" (no files to transfer) sync
run, we would emit the message.  That's wrong.

This patch fixes both problems:
1) it uses the discovered source file count.  If nonzero, we will not
emit the message.

2) it looks at the count of files to delete.  If zero, we will not
emit the message.