Get md5 from remote attribute header in recursive syncs too
We were not getting the md5 value from the attribute header of each
file for the recursive case, only for the non-recursive case.  That
was stupid.  We were certainly _setting_ the md5value into the
attribute header.

This should eliminate the
INFO: disabled md5 check for...
messages on subsequent syncs when s3cmd 1.5.0-alpha+ is used to upload
content in the first place.