FileLists: get md5 from headers
In the recursive case, we really don't care if require_attribs is set,
but we do still care if the existing MD5 value (from ETag) is a
multi-part object.  In that case, we definitely want to try to get it
from the headers.

Drop test for require_attribs.  Keep test for "-" in ETag.