Add --no-check-transfers option
This option deactivates the checking of the MD5 sum of received/uploaded
files. This can be useful if you're downloading/uploading through a
proxy modifying the data on the fly, for example encrypting or
decrypting it.

Since MD5 sums are not checked anymore, there is a risk of getting
corrupted files.