add verbosity when doing local file I/O for stat() and md5 calculation
As this part of the process can take a very long time, especially on
large lists of files to consider, print an info() line at
the start, and then after every 1000 files processed, so the user
knows there's something going on.

This requires the user to run with -v (verbose) turned on to see it,
otherwise it's silent.