Print a proper error for missing dateutil module
Catch the failure to import dateutil where it happens,
not later in s3cmd's ImportError handler.  As this is a new
dependency, many people don't have it installed already.

Without this, the s3cmd ImportError handler (invoked because the
 import in S3/Utils.py of dateutil fails), throws another uncaught
exception when invoking
  s = u' '.join([unicodise(a) for a in sys.argv])
because unicodeise() came from S3/Utils.py which, as just noted,
 failed to import.