add S3/ExitCodes, use them in all sys.exit() calls.
Beginning the process of returning useful exit codes.  First we
add the ExitCodes themselves, and start using them in sys.exit() calls
throughout for consistency.

Some are still a little weird, like returning EX_SOFTWARE when a
request to move a file results in an S3Error(NoSuchKey).  It's not
really a software failure, it's arguably a user input failure.  But to
fix it we have to catch the S3Error lower down, and raise a
ParameterError or similar.

None of the commands (put, get, sync, ...) yet return any status code,
so main() always exits with EX_OK when it runs to completion.