Fix subcmd_batch_del() given empty remote_list
If remote_list was passed, but was empty, the test
   if len([item for item in [uri_str, bucket, remote_list] if item]) != 1

would fail (if remote_list on an empty list returns False), as would
uri_str and bucket being None.  So we would incorrectly fail.

Fix is to catch remote_list being not None and empty before here,
and return gracefully.