Add --ca-cert=FILE option for python 2.7.9ish
Bug #426

python 2.7.9 and 3.x have enabled SSL certificate verification by
default now.  However, when a site uses a self-signed cert, the local
system's CA certificate list likly won't include any self-signed
certs.

We need a way to let the s3cmd caller provide the cert.

This patch allows the caller to specify a path to the public CA cert
file for a given self-signed cert.  Command line option --ca-cert=FILE
is added.

We had to work around the fact that earlier pythons don't have the ssl
or httplib patches that are in 2.7.9 now.  So we test for the
existance of ssl.SSLContext attribute and
httplib.HTTPSConnection(context=) argument, and only use them if present.