Use ca_certs_file="" so it will load from ~/.s3cfg
In ~/.s3cfg, when ca_certs_file = None, it won't get picked up because
it's not of a type the config parser expects.  So, we need it to be ""
instead.

Then, ssl.create_default_context() expects cafile to be either None
(the default), or a string.  So, if we see the empty string, we send
it None instead.