add --remove-headers option, used in [modify]
The [modify] command starts off by making a copy of all the headers on
the source object, then applying them to the destination object.

This patch lets us remove a header from an object using the [modify]
command.  When the copy of the source object headers are made, we then
remove headers from the remove-headers list, then issue the PUT with
the reduced set of headers.