Send requests to best S3 endpoints when --region is known
With the new v4 signature code, we have to send requests to the proper
region.  The default region (US, s3.amazonaws.com) will redirect to
the proper region, but we have to take a redirect to get there.

When --region is specified, and we know that region's endpoint, simply
send the request straight there without being redirected.