Disable ssl hostname checking for all amazonaws.com endpoints
Not just s3.amazonaws.com (US1 default), but for all regions.

If you don't specify --region, it was working, because the initial
request would go to s3.amazonaws.com, and we would then globally
disable hsotname checking.  But if you specify --region eu-central-1,
it would fail because hostname checking would remain on.

This patch checks for presence of 'amazonaws.com' in the hostname, and
disables hostname checking.