remove extraneous call to checksum_sha256()
In object_put() we were calculating the checksum in an argument to
call create_request().  But we were calculating the checksum a second
time in the call to send_file().  For large files, this would invoke
file I/O for a second time, and in any case wasted CPU.

So, we remove the call to checksum_sha256() in the invocation of
create_request() in object_put().