Add http 403 handler for eucalyptus response to signature v4
Eucalyptus walrus doesn't yet support signature v4, it rejects the
request with an HTTP 403.

This patch adds a 403 handler for exactly this returned message, and
reissues the call using signature v2.

N.B. I haven't tested this as I don't have an instance of Walrus
available to me at the moment.