fix S3 wildcard certificate checking
First, the implementation had a bug, and did not actually disable the
ssl.match_hostname() check like we originally believed, because the
change to the context.check_hostname=False flag occurred too late.

Second, we want to do the hostname checking ourselves, so we can
continue when presented with an S3 wildcard certificate.  We invoke
the SSLContext.match_hostname() function for all connections, and if
it fails, we check if we're actually looking at an S3 wildcard
certificate and communicating with an S3 host. If so, this is
permissible.  If not, we fail the connection immediately.

This should resolve the concern over disabling SSL certificate
hostname checking.