Handle headers with spaces in their values properly (#460)
If headers exist to be sent in a request, and the value for the header
contains a space character, the canonical header value being used for
v4 signing would be corrupted - the value was being split at the space
and the two parts of the value were appearing in different parts of
the canonical header after applying a string sort to it.  Not good.

This patch keeps the canonical headers intact in dict format until
generating the final string value of the headers and their values, to
be signed.

It also updates the test case for --add-header to include a space in
the value of the header being added, which would have caught this
failure earlier.

Fixes https://github.com/s3tools/s3cmd/issues/460