Multipart: send 2nd+ chunks from start of buffer
When doing a multipart upload from stdin, we were reading the next
part from stdin into a buffer, and then sending down an offset that
was from the start of the stdin, not from the start of the buffer we
just read into.  Whoops.

This patch fixes it, so we send down offset=0 when sending down a
buffer, so we start from the beginning of the buffer.

This fixes multipart stdin uploads of >1 chunk.