reduce use of os.system() in run-tests stdin tests
Instead of using os.system(), we can open the file that we want to put
via stdin, and pass that file handle to test_s3cmd(), which gets it to
the Popen() call as the stdin argument.  This way, any failure in
s3cmd we can capture and report like for all other tests.