Fix directory exclusions (#467)
Different versions of python append different strings to the pattern
created when turning a glob into a regular expression.  For example,
the pattern:

.snapshot/

on python 2.4, glob.fnmatch.translate() would yield '.snapshot/',
while on python 2.6 and above, we get u'\.snapshot\\/\\Z(?ms)'.

Our test for "is this pattern a directory" was thus failing on python
2.6, testing for 'endswith('/').  Whoops.

This patch tests for both types of pattern endings now.

This fixes #467 and fixes Fedora Infrastructure's sync of EPEL content
into S3 regional mirrors.