Fix indentation causing failures
Commit 41787acd9 accidentally over-indented its change.  This was
leading to test suite failures, valid failures.  Specifically, the
test suite would fail when removing buckets that have objects in them
(the first test of the suite).

Reading that patch, it's obvious that the "if error_node is not None"
test got indented one level too deep from where it was supposed to
be.  Fixing that resolves the test suite failure again.