Catch get_bucket_location IOError in send_file
I frequently see a broken pipe, which I tracked down to get_bucket_location. This seems to happen on big syncs, where there are lots of files to compare - maybe the connection is being dropped?

Borrowing the retry code from the try block below, this simply retries the same send_file, which appears to solve the problem.