escape - and 'Cache-Control' items in manpage
Suggested by Debian patches.  This automatically escapes dashes and
the word 'Cache-Control' in the generated manpage.

There is still an instance where the word Cache-Control is getting
split across newlines due to the 80-character limit, which it doesn't
catch, but it doesn't break the formatting of the rest of the
manpage. It just looks like 'Cache- Control' in the result (note the
extra space after the dash).  I'll live with that.