[info] use last-modified value, not date
The date header contains the remote server's current time.  That's not
what we want.  Instead, we want the last-modified value for the file
being inspected with the [info] (or equivalent) request.

Thanks to Andrew Trusty for finding, diagnosing, and suggesting a
fix.  And thanks to DreamObjects for "forgetting" to include a date
header in their response to a HEAD of an object on occasion, that led
us to find this.

This should also improve --no-check-md5 cases where the timestamp is
being checked, as we won't think that remote files have a timestamp of
"now".