Handle AWS_SESSION_TOKEN environment variable.
This allows s3cmd to be used with temporary credentials as provided by
aws-cli's `aws sts get-session-token` command, which is useful if your
API key is required by a policy to use MFA.

Add option to fetch temporary credentials from EC2's metadata service
even if the config file or environment variables have them.  This can
be convenient if you'd rather not remove environment variables from
your environment but you do want to use an IAM role.

Rename access_token to session_token.  This is the canonical name for
the temporary token, see [0].

[0] http://blogs.aws.amazon.com/security/post/Tx3D6U6WSFGOK2H/A-New-and-Standardized-Way-to-Manage-Credentials-in-the-AWS-SDKs