Add stop_on_error
Add a configuration option that allows an error return on any transfer
error of a file. This is a config option (stop_on_error)  and a
commandline option (--stop-on-error). The Default is the current
continue behavior. With the option or config setting to True, the
commands (get, put, sync remote2remote, remote2local, or local2remote)
will all return an error status. The Get and Put commands return
EX_PARTIAL where the typical response on sync will be EX_TEMPFAIL but
could be any of:

* EX_SERVERMOVED
* EX_SERVERERROR
* EX_ACCESSDENIED
* EX_NOTFOUND
* EX_CONFLICT
* EX_PRECONDITION
* EX_SERVICE
* EX_SOFTWARE

Signed-off-by: Tim Anderson <tsa@biglakesoftware.com>