Fix testsuite for recent changes
1) Use EX_PARTIAL return code in a lot of places.  The unreadable file
was causing return codes of EX_PARTIAL instead of 0 as expected.
2) The string "File: '%s' saved as '%s'" has changed to simply "%s ->
%s" so fix all the expected responses.

This doesn't resolve all errors. There is still one failed return code
renaming a non-existant file, which needs to get tracked down.  That
returns EX_PARTIAL instead of "file not found" as expeted.  But that's
the wrong return code.