introduce class ExitScoreboard
The ExitScoreboard is intended as a helper class to determine if
EX_OK, EX_PARTIAL, EX_NOTFOUND, or EX_GENERAL should be returned.
During a looping execution, calls to
ExitScoreboard.(success,failed,notfound)() updates the scoreboard.  A
final call to rc() calculates the best return code from the scoreboard
results and returns it.

Examples:

EX_OK should be returned when all commands looped over (e.g. a
recursive or wildcard operation) succeeded.

EX_PARTIAL should be returned when some commands looped over
succeeded, while others failed.

EX_NOTFOUND should be returned when all of the items on a list failed
due to not being found.

EX_GENERAL should be returned in any other corner case not accounted
for.

This is only used in subcmd_cp_mv as the testsuite tripped this up
trying to move a nonexistant file.  s3cmd was returning EX_PARTIAL
even though it was a single file being operated on, and it was not
found, so the return code should have been EX_NOTFOUND.