fix md5 checks on download of multipart-uploaded files  (#604)
Bug 604 reported that files [put] from stdin, when later [get], print
a warning that the MD5sums do not match.

Files uploaded via multipart upload (either large files, or files read
from stdin) have ETags on their metadata that include '-'.  These are
known to not be the actual md5.  However, the recv_file() path did not
include code to exclude comparing such md5 values with the calculated
value.

Furthermore, this code path is U-G-L-Y and nearly unreadable.  This
attempts to fix that only slightly, in so far as to correct the logic
for ETags with '-' in them, by re-arranging code blocks, changing
variable names to not be overloaded and be more meaningful, and
skipping the calculation of md5 on the received file (including
possibly reading the file back from the file system) when the ETag
from S3 is used and has that '-' in it, which makes comparison
meaningless.  This should have the positive effect of a) working
correctly; b) easier to read; c) faster for partial downloads (ranges
of a file); d) reducing useless warnings.

This does leave open the possibility of "silent corruption", in that
we are no longer comparing md5s of files uploaded via stdin when that
file is then downloaded.  But we don't have the information to make
the comparison - why bother warning that "gee, we don't know if it
might have been corrupted - probably not though".

One way to mitigate this would be, after doing a multipart stdin
upload, to then do a COPY/MODIFY on the object, storing our calculated
md5 metadata in s3cmd-attrs.  The problem is, that does a COPY within
S3.  Imagine uploading a multi-GB or multi-TB object from stdin, and
then waiting for the following COPY/MODIFY command to complete just to
store the metadata.  I'm not excited about that, and am going to trust
that tcp, http, and https really don't corrupt content on download
very often - mostly through broken pipes - which we'd see an error
from anyhow.