Fix use of scoreboard
Noted in PR review, we were not breaking out of the loop properly.  In
the event of an error, we should not raise the exception again. We
should simply check cfg.stop_on_error, and if set, break out of the
for loop, otherwise keep going.  Once the loop is completed or broken
out of, then return the calcualted return code.