Use test_curl_HEAD() throughout, add --location flag
In some places we were using the test_curl_HEAD() function, in others
we were manually calling it.  Instead, use the helper function
everywhere.

Add --location flag to curl command so that if S3 returns us a 302
redirect (which it may do for its own purposes), we follow it properly.