Allow acl_public = True/False to be in config file
Normally acl_public = None and it's not present, nor read from,
the config file.  If a user wants to set it in the config file though,
we need to be able to read it from the config file.

We also need to _not_ have the lack of command line options
unintentionally set it back to None.

Add special handling for acl_public, when reading in the
config file, and when handling --acl-public and --acl-private options.