Fixes #778 : Workaround for aws sending early errors before requests are completed + support for http expect continue feature.
Added optional support for HTTP Except: 100 Continue feature.
- New option (default False): use_http_except.
For each send_file request, waiting for ok status before sending the data
(Reduce between usage but can introduce a small latency for each send file request
because of the 100-continue status that we wait from the server)
- Aws interrupt early the connection when you try to send a big file and that it
knows in advance that there will be an error (ex. redirect) and that you don't
use the "http expect continue" header. Handle automatically this situation
by trying to read the last http status when an EPIPE or ECONNRESET is encountered.