Fixes #798 : HEAD request can't parse xml response to get redirection url.
So, try to extract this information from the "location" http header that can be provided in such a case.
This don't work with "path style" bucket syntax
(ex.: host_bucket= s3.amazonaws.com and bucket in another region than us-east-1)
because in that case, "location" header is not provided by aws servers.