Py3 support: Reworked request/signature regarding unicode/bytes mix.
Refactored a lot of code to be more clean regarding the mix of "unicode"
and "bytes". Now, bytes should only be used at the most external points
and not be used inside string contructions.

IMPORTANT NOTE: "quote" is now used instead of "quote_plus" everywhere
as aws documentation looks like to have clarified that it is the right
encoding to use. This change may trigger signature issues with some s3
compatible servers I guess.

After this commit, "put files" with exotic characters will work in
python2 and python3 with signature v2 and signature v4.