Fixes #960 - Do a retry if exc has errno but errno is None. Python3 support.
Since py 3, there are new "OS-exception" including RemoteDisconnected.

Sadly, these exception are oserror and are supposed to have an errno
field, but if they are raised by some parts of cpython will not have any
value for errno.

That is the case for RemoteDisconnected raised by httlib instead of
BadStatusLine previously.