Enable AWS cred file to be read as long as .s3cfg auth fields are nonexistent or empty.
This commit allows the credentials to be read from ~/.aws/credentials
or user-specified (via envvar) file if .s3cfg exists but does not have
authentication options.

The previous behavior is that s3cmd will *not* read credentials from the
above path as long as .s3cfg exists (even if it contains no
authentication options).