Fixes #1054 - Relax limitation on special chars for --add-header key names
We are too restrictive on the list of characters that are allowed to be
used in "header name" for the --add-headers option.

The http specification a larger set of characters for header names.
Adding the following ones: !#$%&*+^_|
Still, even if no issue is to be expected, I would not advice anyone to
use any of these chars for a header name.