MultipartCopy - Preserve metadata of the object that is copied
aws s3 does not preserve metadata when doing a copy using
MultipartUploadCopy unlike when doing a simple normal copy.

Sadly, in the case of the multipart copy, that forces us to issue an
additional object_info request as we need to have the original headers
before initiating the MultipartUpload operation.

This patch could still be improved.
A lot earlier in the code, we might have already issued an object_info
request on the source object, but sadly we don't keep the object headers
to re-use them. That could probably be changed to avoid the useless
object_info extra requests.