Fixes #1091 - Ugly hack to have modify work with files bigger than 5Gb
If we want to modify the metadata of a file and that this file is bigger
than the absolute maximum size for a file without multipart(5Gb), then
we run an operation of "Multipart Copy" of this file on itself.
That will probably be slow as it will be like if we do a complete
remote2remote copy of a file, just it will be copied with its own data
at its own path.
Doing that, we can override the metadata that will be used for the "new"
copy.
Really hacky and ugly, but that works...