Fixes #1088 - Handle servers indicating (lying) that listObject result is truncated, when it is not the case
For unknown reasons, it happened with a Dell EMC ECS Object Storage that
a ListObject for a single object returned with the Truncated flag set to
True, despite not being any other result to fetch.
Before this fix, the code was not expecting that and so crashing in that
case.