Take into account the AWS_STS_REGIONAL_ENDPOINTS
This fixes the issue where the sts endpoint sts.amazonaws.com is not
accessible.
Take into account the environment varialbe AWS_STS_REGIONAL_ENDPOINTS
If the variable is provided check for AWS_REGION env var to decide
region.
IF AWS_REGION is not provided use the same region as the bucket.
(bucket_location parameter)