Fixes #996 - Avoid crashing when setup.py test is called
"setup.py test" command is deprecated and should not be used, but still
we are not able to block it.
By default, it can easily crash trying to load python2 or python3 code,
because unittest try to parse all the source files to find tests.

Here we use the "test_suite" arg of setup.py to target an almost empty
module so that we are sure that the command will succeed even if finding
to test to run.