Re-work bucket naming rules.
Add bucket_name_quirks config option (boolean).  This allows you to use a colon
in the bucket name for Ceph compatibility with tenants.

Since it's after March 1, 2018 apply the bucket naming rules for bucket
creation across the board.  Still allow dns_strict to be used if
we ever check bucket names for not creation events.  See

   https://docs.aws.amazon.com/AmazonS3/latest/userguide/bucketnamingrules.html

for details.