Fixes #1256 - Add cfinval command to request cloudfront to invalidate paths
As described in Cloudfront documentation, you can specify the path for
individual files to only invalidate these files.
Or you can specify a path that ends with the * wildcard to request cf to
invalidate all files starting with this path.

If only a bucket is specified, or the specified path ends with "/", a
wildcard is automatically used to invalidate all the files of this
bucket or folder.

To use the wildcard in the command line when calling s3cmd, you must use
quotes (") around the path to avoid your shell processing the "*"
character.